refactor: simplify dev watch and migrate to React 18
