+++ package.json
@@ -5,8 +5,8 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "babel src --out-dir dist --presets=es2015,react",
-    "watch-test": "watchify test/app.js -o test/built.js -t [ babelify --presets [ es2015 react ] ]",
-    "prepublishOnly": "npm run build"
+    "watch-test": "esbuild test/app.js --bundle --outfile=test/built.js --watch",
+    "prepack": "npm run build"
   },
   "repository": {
     "type": "git",
@@ -23,19 +23,13 @@
   },
   "homepage": "https://github.com/cheminfo/visualizer-react#readme",
   "peerDependencies": {
-    "react": "^16.6.0"
+    "react": ">=18.0.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babelify": "^10.0.0",
-    "browserify": "^16.2.3",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "watchify": "^3.11.0"
+    "esbuild": "^0.21.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
-  "dependencies": {},
   "files": [
     "dist"
   ]
